Swapped using an api token for an api key
There seems to be no support/functionality for the api tokens. To circumvent this, I have changed the /get/api_token endpoint to /auth and it will return an api key. The system will first check if the user has an existing api key and if they do, it will be destroyed. Next a new api key is generated and assigned to the user by applying their username as the name for the key. There is still a potential spot in this section to check if the user has 2auth enabled and sending that info back to the client. Will need to add a new endpoint if this is implemented to accept a login attempt using 2auth.