Port change 9090 Added isAdmin check New ApiKeyResponse  class New Bean for connection
The port number for REST requests was changed to 9090. This was done to ensure there would be no fake/accidental requests. A new isAdmin chack was added to the auth endpoint, which then made changing the return setup. This now returns a json in the form of a string, {apiKey, Boolean). This required a new class to make.... maybe not but its done now.